Create abc.py
The first code will print the area of the circle by getting the input as the radius
The second code will search that the string ".py" is in the given string (Input) and print the extension as python. We can extend the code with else if statements for different extensions.